server/python_service/requirements.txt

fastapi
uvicorn
supabase
python-dotenv
pydantic
bcrypt

